Add files via upload
PyTunes.py - The program itself, capable of running either completely on its own OR with the sound effects included. The program will walk you through setup and playing music and whatever. :::NOTE::: THIS PROGRAM NEEDS TO BE IN A DIRECTORY YOU HAVE RW ACCESS TO. IT WILL NOT WORK IN REMOTE DESKTOPS OR FILE SERVERS AND WILL NOT WORK THROUGH WINDOWS 10 QUICK ACCESS.